# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.CreateResultAttachmentRequest:
    properties:
      file_type:
        type: string
      file_url:
        type: string
    type: object
  entities.CreateTestCaseRequest:
    properties:
      description:
        type: string
      expected:
        type: string
      preconditions:
        type: string
      priority:
        type: string
      suite_id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
    type: object
  entities.CreateTestPlanRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  entities.CreateTestResultRequest:
    properties:
      comment:
        type: string
      duration_ms:
        type: integer
      status:
        type: string
      test_case_id:
        type: string
    type: object
...
  entities.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  entities.ResultAttachment:
    properties:
      created_at:
        type: string
      file_type:
        type: string
      file_url:
        type: string
      id:
        type: string
      result_id:
        type: string
    type: object
  entities.TestCase:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      expected:
        type: string
      id:
        type: string
      preconditions:
        type: string
      priority:
        type: string
      project_id:
        type: string
      status:
        type: string
      steps:
        type: string
      suite_id:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  entities.TestCaseHistory:
    properties:
      changed_at:
        type: string
      changed_by:
        type: string
      id:
        type: string
      snapshot: {}
      test_case_id:
        type: string
    type: object
  entities.TestPlan:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
    type: object
  entities.TestResult:
    properties:
      comment:
        type: string
      created_at:
        type: string
      duration_ms:
        type: integer
      executed_by:
        type: string
      id:
        type: string
      run_id:
        type: string
      status:
        type: string
      test_case_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Plain-TCMS API
  version: "1.0"
paths:
...
  /projects/{id}/test-cases:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.TestCase'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test cases by project id
      tags:
      - Test Cases
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Case
        in: body
        name: test_case
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestCaseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.TestCase'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Create test case
      tags:
      - Test Cases
  /projects/{id}/test-plans:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.TestPlan'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test plans by project
      tags:
      - Test Plans
    post:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Plan
        in: body
        name: test_plan
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestPlanRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.TestPlan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Create test plan
      tags:
      - Test Plans
...
  /test-cases/{id}:
    delete:
      parameters:
      - description: Test Case ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete test case
      tags:
      - Test Cases
    get:
      parameters:
      - description: Test Case ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestCase'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test case by ID
      tags:
      - Test Cases
    put:
      consumes:
      - application/json
      parameters:
      - description: Test Case ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Case
        in: body
        name: test_case
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestCaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestCase'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Update test case
      tags:
      - Test Cases
  /test-cases/{id}/history:
    get:
      parameters:
      - description: Test Case ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.TestCaseHistory'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test case history
      tags:
      - Test Cases
  /test-plans/{id}:
    delete:
      parameters:
      - description: Test Plan ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete test plan
      tags:
      - Test Plans
    get:
      parameters:
      - description: Test Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestPlan'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test plan by ID
      tags:
      - Test Plans
    put:
      consumes:
      - application/json
      parameters:
      - description: Test Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Plan
        in: body
        name: test_plan
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestPlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestPlan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Update test plan
      tags:
      - Test Plans
  /test-results/{id}:
    get:
      parameters:
      - description: Test Result ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get test result by ID
      tags:
      - Test Results
    put:
      consumes:
      - application/json
      parameters:
      - description: Test Result ID
        in: path
        name: id
        required: true
        type: string
      - description: Test Result
        in: body
        name: result
        required: true
        schema:
          $ref: '#/definitions/entities.CreateTestResultRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.TestResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Update test result
      tags:
      - Test Results
  /test-results/{id}/attachments:
    get:
      parameters:
      - description: Test Result ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.ResultAttachment'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get attachments for a test result
      tags:
      - Test Results
    post:
      consumes:
      - application/json
      parameters:
      - description: Test Result ID
        in: path
        name: id
        required: true
        type: string
      - description: Attachment
        in: body
        name: attachment
        required: true
        schema:
          $ref: '#/definitions/entities.CreateResultAttachmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.ResultAttachment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Add attachment to test result
      tags:
      - Test Results
...
swagger: "2.0"